Add explicit (size_t) cast for long offset in ZSTDSeek_seek()
The second getJumpCoordinate call passed offset (type long) without a
cast, while the first call at line 622 already used (size_t)offset.
The implicit conversion is well-defined when offset >= 0 (guaranteed
by the earlier check), but the inconsistency can trigger
-Wsign-conversion warnings in strict CI builds.

+++ zstd-seek.c
@@ -629,7 +629,7 @@ int ZSTDSeek_seek(ZSTDSeek_Context *sctx, long offset, int origin){
             return 0;
         }
 
-        const ZSTDSeek_JumpCoordinate new_jc = ZSTDSeek_getJumpCoordinate(sctx, offset);
+        const ZSTDSeek_JumpCoordinate new_jc = ZSTDSeek_getJumpCoordinate(sctx, (size_t)offset);
 
         if(sctx->jc.compressedOffset != new_jc.compressedOffset || offset < sctx->currentUncompressedPos){ //reset
             ZSTD_DCtx_reset(sctx->dctx, ZSTD_reset_session_only);
